+++ packages/fuels-core/src/types/transaction_builders.rs
@@ -12,9 +12,9 @@ use fuel_crypto::{Hasher, Message as CryptoMessage, Signature};
 use fuel_tx::{
     field::{Outputs, Policies as PoliciesField, ScriptGasLimit, Witnesses},
     policies::{Policies, PolicyType},
-    BlobId, BlobIdExt, Chargeable, ConsensusParameters, Create, Input as FuelInput, Output, Script,
-    StorageSlot, Transaction as FuelTransaction, TransactionFee, TxPointer, UniqueIdentifier,
-    Upgrade, Upload, UploadBody, Witness,
+    Chargeable, ConsensusParameters, Create, Input as FuelInput, Output, Script, StorageSlot,
+    Transaction as FuelTransaction, TransactionFee, TxPointer, UniqueIdentifier, Upgrade, Upload,
+    UploadBody, Witness,
 };
 pub use fuel_tx::{UpgradePurpose, UploadSubsection};
 use fuel_types::{bytes::padded_len_usize, Bytes32, Salt};
@@ -40,9 +40,10 @@ use crate::{
     utils::{calculate_witnesses_size, sealed},
 };
 
-use super::transaction::BlobTransaction;
-
+mod blob;
 mod script_tx_estimator;
+
+pub use blob::*;
 
 #[derive(Debug, Clone, Default)]
 struct UnresolvedWitnessIndexes {
@@ -173,10 +174,10 @@ pub trait TransactionBuilder: BuildableTransaction + Send + sealed::Sealed {
     fn with_estimation_horizon(self, block_horizon: u32) -> Self;
 }
 
-macro_rules! impl_tx_trait {
+macro_rules! impl_tx_builder_trait {
     ($ty: ty, $tx_ty: ident) => {
         #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-        impl TransactionBuilder for $ty {
+        impl $crate::types::transaction_builders::TransactionBuilder for $ty {
             type TxType = $tx_ty;
 
             fn add_signer(&mut self, signer: impl Signer + Send + Sync) -> Result<&mut Self> {
@@ -213,7 +214,11 @@ macro_rules! impl_tx_trait {
                     .witnesses_mut()
                     .extend(repeat(witness).take(self.unresolved_signers.len()));
 
-                let mut tx = BuildableTransaction::build(fee_estimation_tb, &provider).await?;
+                let mut tx = $crate::types::transaction_builders::BuildableTransaction::build(
+                    fee_estimation_tb,
+                    &provider,
+                )
+                .await?;
 
                 if tx.is_using_predicates() {
                     tx.estimate_predicates(&provider, None).await?;
@@ -225,7 +230,7 @@ macro_rules! impl_tx_trait {
                     .estimate_gas_price(self.gas_price_estimation_block_horizon)
                     .await?;
 
-                estimate_max_fee_w_tolerance(
+                $crate::types::transaction_builders::estimate_max_fee_w_tolerance(
                     tx.tx,
                     self.max_fee_estimation_tolerance,
                     gas_price,
@@ -287,6 +292,7 @@ macro_rules! impl_tx_trait {
 
         impl $ty {
             fn set_witness_indexes(&mut self) {
+                use $crate::types::transaction_builders::TransactionBuilder;
                 self.unresolved_witness_indexes.owner_to_idx_offset = self
                     .inputs()
                     .iter()
@@ -317,12 +323,14 @@ macro_rules! impl_tx_trait {
             }
 
             fn is_using_predicates(&self) -> bool {
+                use $crate::types::transaction_builders::TransactionBuilder;
                 self.inputs()
                     .iter()
                     .any(|input| matches!(input, Input::ResourcePredicate { .. }))
             }
 
             fn num_witnesses(&self) -> Result<u16> {
+                use $crate::types::transaction_builders::TransactionBuilder;
                 let num_witnesses = self.witnesses().len();
 
                 if num_witnesses + self.unresolved_signers.len() > u16::MAX as usize {
@@ -361,7 +369,7 @@ macro_rules! impl_tx_trait {
                 let gas_price = provider.estimate_gas_price(block_horizon).await?;
                 let consensus_parameters = provider.consensus_parameters();
 
-                let max_fee = estimate_max_fee_w_tolerance(
+                let max_fee = $crate::types::transaction_builders::estimate_max_fee_w_tolerance(
                     wrapper_tx.tx,
                     max_fee_estimation_tolerance,
                     gas_price,
@@ -376,7 +384,9 @@ macro_rules! impl_tx_trait {
     };
 }
 
-fn estimate_max_fee_w_tolerance<T: Chargeable>(
+pub(crate) use impl_tx_builder_trait;
+
+pub(crate) fn estimate_max_fee_w_tolerance<T: Chargeable>(
     tx: T,
     tolerance: f32,
     gas_price: u64,
@@ -524,142 +534,10 @@ impl Default for UpgradeTransactionBuilder {
     }
 }
 
-#[derive(Default, Clone, Debug, PartialEq)]
-pub struct Blob {
-    pub data: Vec<u8>,
-}
-
-impl Blob {
-    pub fn id(&self) -> [u8; 32] {
-        BlobId::compute(&self.data).into()
-    }
-
-    fn as_blob_body(&self, witness_index: u16) -> fuel_tx::BlobBody {
-        fuel_tx::BlobBody {
-            id: self.id().into(),
-            witness_index,
-        }
-    }
-}
-
-impl From<Blob> for fuel_tx::Witness {
-    fn from(blob: Blob) -> Self {
-        blob.data.into()
-    }
-}
-
-#[derive(Default)]
-pub struct BlobTransactionBuilder {
-    pub inputs: Vec<Input>,
-    pub outputs: Vec<Output>,
-    pub witnesses: Vec<Witness>,
-    pub tx_policies: TxPolicies,
-    pub gas_price_estimation_block_horizon: u32,
-    pub max_fee_estimation_tolerance: f32,
-    pub build_strategy: Strategy,
-    pub blob: Blob,
-    unresolved_witness_indexes: UnresolvedWitnessIndexes,
-    unresolved_signers: Vec<Box<dyn Signer + Send + Sync>>,
-}
-
-impl BlobTransactionBuilder {
-    pub async fn build(mut self, provider: impl DryRunner) -> Result<BlobTransaction> {
-        let is_using_predicates = self.is_using_predicates();
-
-        let tx = match self.build_strategy {
-            Strategy::Complete => self.resolve_fuel_tx(&provider).await?,
-            Strategy::NoSignatures => {
-                self.set_witness_indexes();
-                self.unresolved_signers = Default::default();
-                self.resolve_fuel_tx(&provider).await?
-            }
-        };
-
-        Ok(BlobTransaction {
-            is_using_predicates,
-            tx,
-        })
-    }
-
-    fn clone_without_signers(&self) -> Self {
-        Self {
-            inputs: self.inputs.clone(),
-            outputs: self.outputs.clone(),
-            witnesses: self.witnesses.clone(),
-            tx_policies: self.tx_policies,
-            unresolved_witness_indexes: self.unresolved_witness_indexes.clone(),
-            unresolved_signers: Default::default(),
-            gas_price_estimation_block_horizon: self.gas_price_estimation_block_horizon,
-            max_fee_estimation_tolerance: self.max_fee_estimation_tolerance,
-            build_strategy: self.build_strategy.clone(),
-            blob: self.blob.clone(),
-        }
-    }
-
-    async fn resolve_fuel_tx(mut self, provider: &impl DryRunner) -> Result<fuel_tx::Blob> {
-        let chain_id = provider.consensus_parameters().chain_id();
-
-        // TODO: segfault
-        let blob_witness_index = self.witnesses.len() as u16;
-        let body = self.blob.as_blob_body(blob_witness_index);
-        let blob_witness = std::mem::take(&mut self.blob).into();
-        self.witnesses.push(blob_witness);
-
-        let num_witnesses = self.num_witnesses()?;
-        let policies = self.generate_fuel_policies()?;
-        let is_using_predicates = self.is_using_predicates();
-
-        let mut tx = FuelTransaction::blob(
-            body,
-            policies,
-            resolve_fuel_inputs(self.inputs, num_witnesses, &self.unresolved_witness_indexes)?,
-            self.outputs,
-            self.witnesses,
-        );
-
-        if let Some(max_fee) = self.tx_policies.max_fee() {
-            tx.policies_mut().set(PolicyType::MaxFee, Some(max_fee));
-        } else {
-            Self::set_max_fee_policy(
-                &mut tx,
-                &provider,
-                self.gas_price_estimation_block_horizon,
-                is_using_predicates,
-                self.max_fee_estimation_tolerance,
-            )
-            .await?;
-        }
-
-        let signatures =
-            generate_missing_witnesses(tx.id(&chain_id), &self.unresolved_signers).await?;
-        tx.witnesses_mut().extend(signatures);
-
-        Ok(tx)
-    }
-}
-
-impl sealed::Sealed for BlobTransactionBuilder {}
-
-#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-impl BuildableTransaction for BlobTransactionBuilder {
-    type TxType = BlobTransaction;
-    type Strategy = Strategy;
-
-    fn with_build_strategy(mut self, strategy: Self::Strategy) -> Self {
-        self.build_strategy = strategy;
-        self
-    }
-
-    async fn build(self, provider: impl DryRunner) -> Result<Self::TxType> {
-        BlobTransactionBuilder::build(self, provider).await
-    }
-}
-
-impl_tx_trait!(ScriptTransactionBuilder, ScriptTransaction);
-impl_tx_trait!(CreateTransactionBuilder, CreateTransaction);
-impl_tx_trait!(UploadTransactionBuilder, UploadTransaction);
-impl_tx_trait!(UpgradeTransactionBuilder, UpgradeTransaction);
-impl_tx_trait!(BlobTransactionBuilder, BlobTransaction);
+impl_tx_builder_trait!(ScriptTransactionBuilder, ScriptTransaction);
+impl_tx_builder_trait!(CreateTransactionBuilder, CreateTransaction);
+impl_tx_builder_trait!(UploadTransactionBuilder, UploadTransaction);
+impl_tx_builder_trait!(UpgradeTransactionBuilder, UpgradeTransaction);
 
 impl ScriptTransactionBuilder {
     async fn build(mut self, provider: impl DryRunner) -> Result<ScriptTransaction> {
